+++ Gopkg.lock
@@ -1,0 +1,111 @@
+# This file is autogenerated, do not edit; changes may be undone by the next 'dep ensure'.
+
+
+[[projects]]
+  digest = "1:9752dad5e89cd779096bf2477a4ded16bea7ac62de453c8d6b4bf841d51a8512"
+  name = "git.apache.org/thrift.git"
+  packages = ["lib/go/thrift"]
+  pruneopts = "UT"
+  revision = "540e346d6ec3be9893ff1f29ef89e000c314edf4"
+
+[[projects]]
+  digest = "1:e6f4136a59289e23b085e4f7abb2e75eb34e7be3dd82acf0f9592c521224b63e"
+  name = "github.com/BurntSushi/toml"
+  packages = ["."]
+  pruneopts = "UT"
+  revision = "99064174e013895bbd9b025c31100bd1d9b590ca"
+
+[[projects]]
+  digest = "1:c79fb010be38a59d657c48c6ba1d003a8aa651fa56b579d959d74573b7dff8e1"
+  name = "github.com/gorilla/context"
+  packages = ["."]
+  pruneopts = "UT"
+  revision = "08b5f424b9271eedf6f9f0ce86cb9396ed337a42"
+  version = "v1.1.1"
+
+[[projects]]
+  digest = "1:708ca7f70df972672d96e0a205bf6b2025cb35872d04a250c65cdb9632a851dc"
+  name = "github.com/gorilla/mux"
+  packages = ["."]
+  pruneopts = "UT"
+  revision = "757bef944d0f21880861c2dd9c871ca543023cba"
+
+[[projects]]
+  digest = "1:d7299d693d72fd124da14df7d15e1231eea9fd1cc92fbc267b1df30b09669800"
+  name = "github.com/spacemonkeygo/errors"
+  packages = [
+    ".",
+    "errhttp",
+  ]
+  pruneopts = "UT"
+  revision = "17af3f6f14b649913cfaa87d167574d970742f96"
+
+[[projects]]
+  digest = "1:c5d60a4dfca98309e65c140b6c746994d0b23d6c6e94f79f90972e871dead09d"
+  name = "github.com/spacemonkeygo/monotime"
+  packages = [
+    ".",
+    "_cgo",
+  ]
+  pruneopts = "UT"
+  revision = "cdd56516a92a4dc680379a26524c9c3d59d79ec0"
+
+[[projects]]
+  digest = "1:041166c169a282ba0a651bf107ac5395fef8b98fdca38cb3e4ed4cac15761c56"
+  name = "golang.org/x/net"
+  packages = ["context"]
+  pruneopts = "UT"
+  revision = "8fd7f25955530b92e73e9e1932a41b522b22ccd9"
+
+[[projects]]
+  digest = "1:5052389c6c809eeced44ec24ff9958bfea70b9859b7a25061c6fd9086c8f8989"
+  name = "gopkg.in/mgo.v2"
+  packages = [
+    ".",
+    "bson",
+    "internal/json",
+    "internal/sasl",
+    "internal/scram",
+  ]
+  pruneopts = "UT"
+  revision = "3f83fa5005286a7fe593b055f0d7771a7dce4655"
+
+[[projects]]
+  branch = "v2"
+  digest = "1:03e7f5627ac0d18fd9705d88444d56ef9fa3e0c390df662b860e4c522ddc0072"
+  name = "gopkg.in/spacemonkeygo/monkit-zipkin.v2"
+  packages = [
+    ".",
+    "gen-go/scribe",
+    "gen-go/zipkin",
+  ]
+  pruneopts = "UT"
+  revision = "4ccc465eff26f74b6cd0c4b9307f943d4bc34b2e"
+
+[[projects]]
+  digest = "1:1e6f89f89be85f186f2962f87e3e368a7ae1bf5c2718dec0ba1884480058c91c"
+  name = "gopkg.in/spacemonkeygo/monkit.v2"
+  packages = [
+    ".",
+    "collect",
+    "environment",
+    "present",
+  ]
+  pruneopts = "UT"
+  revision = "1eba693673c928f81d8adb80ee893985c0c9c412"
+
+[solve-meta]
+  analyzer-name = "dep"
+  analyzer-version = 1
+  input-imports = [
+    "github.com/BurntSushi/toml",
+    "github.com/gorilla/mux",
+    "gopkg.in/mgo.v2",
+    "gopkg.in/mgo.v2/bson",
+    "gopkg.in/spacemonkeygo/monkit-zipkin.v2",
+    "gopkg.in/spacemonkeygo/monkit.v2",
+    "gopkg.in/spacemonkeygo/monkit.v2/environment",
+    "gopkg.in/spacemonkeygo/monkit.v2/present",
+  ]
+  solver-name = "gps-cdcl"
+  solver-version = 1
