+++ build.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 echo "Building the http-server binary"
-govendor sync
+dep ensure -v
 
 go install github.com/alecholmez/http-server/h2tp
 echo "Done"
