+++ buildDocker.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+echo "Building the http-server docker container"
+docker build -t alecholmez/http-server:latest .
+echo "Done"
