+++ h2tp/main.go
@@ -7,22 +7,24 @@ import (
 	"time"
 
 	"github.com/alecholmez/http-server/config"
+	"github.com/alecholmez/http-server/metrics"
 )
 
 func main() {
 	// Load in the config file and define the environment
 	c := config.NewConfig("../config.toml")
 
-	Setup()
-	Start("0.0.0.0", 9000)
+	metrics.Setup()
+	metrics.Start("0.0.0.0", 9000)
 
+	// Start our zipkin go routine
 	go func() {
 		// Allow time for zipkin to start up but don't block the process
 		log.Println("sleeping for 10 seconds...")
 		time.Sleep(10 * time.Second)
 
 		log.Println("Starting zipkin registration")
-		if err := RegisterZipkin("zipkin", 9411); err != nil {
+		if err := metrics.RegisterZipkin("zipkin", 9411); err != nil {
 			log.Fatal(err)
 		}
 		log.Println("Finished zipkin registration")
@@ -35,7 +37,7 @@ func main() {
 	}
 
 	// Create the middleware stack
-	stack := NewStack(RTS, c)
+	stack := NewStack(CreateRoutes(), c)
 
 	// Listen and serve using the middleware stack
 	log.Printf("Server is locally listening on port %s\n", port)
